Delete ChapterDTO & use getMangaUrl
