--- src/vi/mimihentai/src/eu/kanade/tachiyomi/extension/vi/mimihentai/ChapterDTO.kt
@@ -1,10 +1,0 @@
-package eu.kanade.tachiyomi.extension.vi.mimihentai
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-class ChapterDTO(
-    val id: Long,
-    val title: String,
-    val createdAt: String,
-)
